ws-gateway/pyproject.toml
[project]
name = "ws-gateway"
version = "0.1.0"
description = "WebSocket gateway for Chatodon"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "httpx>=0.28.1",
    "redis>=7.3.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = ["ruff>=0.15.6"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle
    "F",  # pyflakes
    "I",  # isort
    "UP", # pyupgrade
    "B",  # bugbear
]

ignore = [
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"migrations/*.py" = ["E501", "F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
